requirements.txt
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-httpx